Refactor MarkdownCombinerPlugin to use CombinerApp component; remove unused MarkdownProcessor and PluginMain components; implement markdown file combining functionality with embedded links processing.
