+++ main.js
@@ -3,7 +3,7 @@ THIS IS A GENERATED/BUNDLED FILE BY ESBUILD
 if you want to view the source, please visit the github repository of this plugin
 */
 
-"use strict";var Y=Object.defineProperty;var Dt=Object.getOwnPropertyDescriptor;var Tt=Object.getOwnPropertyNames;var jt=Object.prototype.hasOwnProperty;var Lt=(t,e)=>{for(var o in e)Y(t,o,{get:e[o],enumerable:!0})},Pt=(t,e,o,r)=>{if(e&&typeof e=="object"||typeof e=="function")for(let n of Tt(e))!jt.call(t,n)&&n!==o&&Y(t,n,{get:()=>e[n],enumerable:!(r=Dt(e,n))||r.enumerable});return t};var Rt=t=>Pt(Y({},"__esModule",{value:!0}),t);var ve={};Lt(ve,{default:()=>K});module.exports=Rt(ve);var Ot=require("obsidian");function k(){}function It(t,e){for(let o in e)t[o]=e[o];return t}function Q(t){return t()}function R(){return Object.create(null)}function x(t){t.forEach(Q)}function I(t){return typeof t=="function"}function q(t,e){return t!=t?e==e:t!==e||t&&typeof t=="object"||typeof t=="function"}function ut(t){return Object.keys(t).length===0}function B(t,e,o,r){if(t){let n=at(t,e,o,r);return t[0](n)}}function at(t,e,o,r){return t[1]&&r?It(o.ctx.slice(),t[1](r(e))):o.ctx}function z(t,e,o,r){if(t[2]&&r){let n=t[2](r(o));if(e.dirty===void 0)return n;if(typeof n=="object"){let i=[],s=Math.max(e.dirty.length,n.length);for(let l=0;l<s;l+=1)i[l]=e.dirty[l]|n[l];return i}return e.dirty|n}return e.dirty}function U(t,e,o,r,n,i){if(n){let s=at(e,o,r,i);t.p(s,n)}}function G(t){if(t.ctx.length>32){let e=[],o=t.ctx.length/32;for(let r=0;r<o;r++)e[r]=-1;return e}return-1}var Z=typeof window<"u"?window:typeof globalThis<"u"?globalThis:global;var W=class t{_listeners="WeakMap"in Z?new WeakMap:void 0;_observer=void 0;options;constructor(e){this.options=e}observe(e,o){return this._listeners.set(e,o),this._getObserver().observe(e,this.options),()=>{this._listeners.delete(e),this._observer.unobserve(e)}}_getObserver(){return this._observer??(this._observer=new ResizeObserver(e=>{for(let o of e)t.entries.set(o.target,o),this._listeners.get(o.target)?.(o)}))}};W.entries="WeakMap"in Z?new WeakMap:void 0;var ft=!1;function pt(){ft=!0}function dt(){ft=!1}function _t(t,e){t.appendChild(e)}function H(t,e,o){let r=ht(t);if(!r.getElementById(e)){let n=E("style");n.id=e,n.textContent=o,zt(r,n)}}function ht(t){if(!t)return document;let e=t.getRootNode?t.getRootNode():t.ownerDocument;return e&&e.host?e:t.ownerDocument}function zt(t,e){return _t(t.head||t,e),e.sheet}function v(t,e,o){t.insertBefore(e,o||null)}function $(t){t.parentNode&&t.parentNode.removeChild(t)}function E(t){return document.createElement(t)}function Ut(t){return document.createTextNode(t)}function mt(){return Ut("")}function S(t,e,o){o==null?t.removeAttribute(e):t.getAttribute(e)!==o&&t.setAttribute(e,o)}function xt(t){return Array.from(t.childNodes)}function $t(t){let e={};return t.childNodes.forEach(o=>{e[o.slot||"default"]=!0}),e}var A;function w(t){A=t}var C=[];var yt=[],D=[],bt=[],Xt=Promise.resolve(),et=!1;function wt(){et||(et=!0,Xt.then(V))}function L(t){D.push(t)}var tt=new Set,N=0;function V(){if(N!==0)return;let t=A;do{try{for(;N<C.length;){let e=C[N];N++,w(e),Kt(e.$$)}}catch(e){throw C.length=0,N=0,e}for(w(null),C.length=0,N=0;yt.length;)yt.pop()();for(let e=0;e<D.length;e+=1){let o=D[e];tt.has(o)||(tt.add(o),o())}D.length=0}while(C.length);for(;bt.length;)bt.pop()();et=!1,tt.clear(),w(t)}function Kt(t){if(t.fragment!==null){t.update(),x(t.before_update);let e=t.dirty;t.dirty=[-1],t.fragment&&t.fragment.p(t.ctx,e),t.after_update.forEach(L)}}function vt(t){let e=[],o=[];D.forEach(r=>t.indexOf(r)===-1?e.push(r):o.push(r)),o.forEach(r=>r()),D=e}var J=new Set,M;function nt(){M={r:0,c:[],p:M}}function ot(){M.r||x(M.c),M=M.p}function m(t,e){t&&t.i&&(J.delete(t),t.i(e))}function F(t,e,o,r){if(t&&t.o){if(J.has(t))return;J.add(t),M.c.push(()=>{J.delete(t),r&&(o&&t.d(1),r())}),t.o(e)}else r&&r()}var Yt=["allowfullscreen","allowpaymentrequest","async","autofocus","autoplay","checked","controls","default","defer","disabled","formnovalidate","hidden","inert","ismap","loop","multiple","muted","nomodule","novalidate","open","playsinline","readonly","required","reversed","selected"],Qt=new Set([...Yt]);function ee(t,e,o){let{fragment:r,after_update:n}=t.$$;r&&r.m(e,o),L(()=>{let i=t.$$.on_mount.map(Q).filter(I);t.$$.on_destroy?t.$$.on_destroy.push(...i):x(i),t.$$.on_mount=[]}),n.forEach(L)}function ne(t,e){let o=t.$$;o.fragment!==null&&(vt(o.after_update),x(o.on_destroy),o.fragment&&o.fragment.d(e),o.on_destroy=o.fragment=null,o.ctx=[])}function oe(t,e){t.$$.dirty[0]===-1&&(C.push(t),wt(),t.$$.dirty.fill(0)),t.$$.dirty[e/31|0]|=1<<e%31}function X(t,e,o,r,n,i,s=null,l=[-1]){let f=A;w(t);let c=t.$$={fragment:null,ctx:[],props:i,update:k,not_equal:n,bound:R(),on_mount:[],on_destroy:[],on_disconnect:[],before_update:[],after_update:[],context:new Map(e.context||(f?f.$$.context:[])),callbacks:R(),dirty:l,skip_bound:!1,root:e.target||f.$$.root};s&&s(c.root);let a=!1;if(c.ctx=o?o(t,e.props||{},(d,u,...y)=>{let p=y.length?y[0]:u;return c.ctx&&n(c.ctx[d],c.ctx[d]=p)&&(!c.skip_bound&&c.bound[d]&&c.bound[d](p),a&&oe(t,d)),u}):[],c.update(),a=!0,x(c.before_update),c.fragment=r?r(c.ctx):!1,e.target){if(e.hydrate){pt();let d=xt(e.target);c.fragment&&c.fragment.l(d),d.forEach($)}else c.fragment&&c.fragment.c();e.intro&&m(t.$$.fragment),ee(t,e.target,e.anchor),dt(),V()}w(f)}var re;typeof HTMLElement=="function"&&(re=class extends HTMLElement{$$ctor;$$s;$$c;$$cn=!1;$$d={};$$r=!1;$$p_d={};$$l={};$$l_u=new Map;constructor(t,e,o){super(),this.$$ctor=t,this.$$s=e,o&&this.attachShadow({mode:"open"})}addEventListener(t,e,o){if(this.$$l[t]=this.$$l[t]||[],this.$$l[t].push(e),this.$$c){let r=this.$$c.$on(t,e);this.$$l_u.set(e,r)}super.addEventListener(t,e,o)}removeEventListener(t,e,o){if(super.removeEventListener(t,e,o),this.$$c){let r=this.$$l_u.get(e);r&&(r(),this.$$l_u.delete(e))}if(this.$$l[t]){let r=this.$$l[t].indexOf(e);r>=0&&this.$$l[t].splice(r,1)}}async connectedCallback(){if(this.$$cn=!0,!this.$$c){let t=function(n){return()=>{let i;return{c:function(){i=E("slot"),n!=="default"&&S(i,"name",n)},m:function(f,c){v(f,i,c)},d:function(f){f&&$(i)}}}};if(await Promise.resolve(),!this.$$cn||this.$$c)return;let e={},o=$t(this);for(let n of this.$$s)n in o&&(e[n]=[t(n)]);for(let n of this.attributes){let i=this.$$g_p(n.name);i in this.$$d||(this.$$d[i]=rt(i,n.value,this.$$p_d,"toProp"))}for(let n in this.$$p_d)!(n in this.$$d)&&this[n]!==void 0&&(this.$$d[n]=this[n],delete this[n]);this.$$c=new this.$$ctor({target:this.shadowRoot||this,props:{...this.$$d,$$slots:e,$$scope:{ctx:[]}}});let r=()=>{this.$$r=!0;for(let n in this.$$p_d)if(this.$$d[n]=this.$$c.$$.ctx[this.$$c.$$.props[n]],this.$$p_d[n].reflect){let i=rt(n,this.$$d[n],this.$$p_d,"toAttribute");i==null?this.removeAttribute(this.$$p_d[n].attribute||n):this.setAttribute(this.$$p_d[n].attribute||n,i)}this.$$r=!1};this.$$c.$$.after_update.push(r),r();for(let n in this.$$l)for(let i of this.$$l[n]){let s=this.$$c.$on(n,i);this.$$l_u.set(i,s)}this.$$l={}}}attributeChangedCallback(t,e,o){this.$$r||(t=this.$$g_p(t),this.$$d[t]=rt(t,o,this.$$p_d,"toProp"),this.$$c?.$set({[t]:this.$$d[t]}))}disconnectedCallback(){this.$$cn=!1,Promise.resolve().then(()=>{!this.$$cn&&this.$$c&&(this.$$c.$destroy(),this.$$c=void 0)})}$$g_p(t){return Object.keys(this.$$p_d).find(e=>this.$$p_d[e].attribute===t||!this.$$p_d[e].attribute&&e.toLowerCase()===t)||t}});function rt(t,e,o,r){let n=o[t]?.type;if(e=n==="Boolean"&&typeof e!="boolean"?e!=null:e,!r||!o[t])return e;if(r==="toAttribute")switch(n){case"Object":case"Array":return e==null?null:JSON.stringify(e);case"Boolean":return e?"":null;case"Number":return e??null;default:return e}else switch(n){case"Object":case"Array":return e&&JSON.parse(e);case"Boolean":return e;case"Number":return e!=null?+e:e;default:return e}}var O=class{$$=void 0;$$set=void 0;$destroy(){ne(this,1),this.$destroy=k}$on(e,o){if(!I(o))return k;let r=this.$$.callbacks[e]||(this.$$.callbacks[e]=[]);return r.push(o),()=>{let n=r.indexOf(o);n!==-1&&r.splice(n,1)}}$set(e){this.$$set&&!ut(e)&&(this.$$.skip_bound=!0,this.$$set(e),this.$$.skip_bound=!1)}};var Ft="4";typeof window<"u"&&(window.__svelte||(window.__svelte={v:new Set})).v.add(Ft);var g=require("obsidian");var St=require("obsidian");function _e(t){H(t,"svelte-9ekks4",".markdown-processor.svelte-9ekks4{width:100%}")}var he=t=>({processedContent:t&1}),kt=t=>({processedContent:t[0]});function Et(t){let e,o,r=t[6].default,n=B(r,t,t[5],kt);return{c(){e=E("div"),n&&n.c(),S(e,"class","markdown-processor svelte-9ekks4")},m(i,s){v(i,e,s),n&&n.m(e,null),o=!0},p(i,s){n&&n.p&&(!o||s&33)&&U(n,r,i,i[5],o?z(r,i[5],s,he):G(i[5]),kt)},i(i){o||(m(n,i),o=!0)},o(i){F(n,i),o=!1},d(i){i&&$(e),n&&n.d(i)}}}function me(t){let e,o,r=t[0]&&Et(t);return{c(){r&&r.c(),e=mt()},m(n,i){r&&r.m(n,i),v(n,e,i),o=!0},p(n,[i]){n[0]?r?(r.p(n,i),i&1&&m(r,1)):(r=Et(n),r.c(),m(r,1),r.m(e.parentNode,e)):r&&(nt(),F(r,1,1,()=>{r=null}),ot())},i(n){o||(m(r),o=!0)},o(n){F(r),o=!1},d(n){n&&$(e),r&&r.d(n)}}}function xe(t,e){let o=t.split(`
-`),r=new RegExp(`^#+\\s*${e}$`),n=!1,i=[],s=0;for(let l of o){let f=l.match(/^(#+)\s*(.*)$/);if(f){let c=f[1].length;r.test(l)?(n=!0,s=c,i.push(l)):n&&c<=s&&(n=!1)}n&&i.push(l)}return i.join(`
-`)}function $e(t,e){let o=t.split(`
-`),r=new RegExp(`\\s\\^${e}$`);for(let n=0;n<o.length;n++)if(r.test(o[n]))return o[n].replace(r,"").trim();return""}function ge(t,e,o){let{$$slots:r={},$$scope:n}=e,{app:i}=e,{content:s=""}=e,{processedContent:l=""}=e,f=new Set;async function c(u){let y=/!\[\[([^\]#|^]+)(?:(#)([^\]|]+))?(?:(\^)([^\]|]+))?(?:\|([^\]]+))?\]\]/g,p=u,_;for(;(_=y.exec(u))!==null;){let b=_[0],h=_[1].trim(),T=_[2],ct=_[3],At=_[4],lt=_[5];if(f.has(h)){console.warn(`Circular reference detected, skipping: ${h}`);continue}f.add(h);try{let P=i.metadataCache.getFirstLinkpathDest(h,"");if(P instanceof St.TFile){let j=await i.vault.read(P);T&&ct?j=xe(j,ct):At&&lt&&(j=$e(j,lt));let Nt=await c(j);p=p.replace(b,Nt)}else p=p.replace(b,`<!-- Linked file not found: ${h} -->`)}catch(P){console.error(`Error processing link ${b}:`,P),p=p.replace(b,`<!-- Error: Could not process ${h} -->`)}finally{f.delete(h)}}return p}async function a(u){return f.clear(),await c(u)}async function d(u,y){let p=1,_=`${u}-combined.md`;for(;await i.vault.adapter.exists(`${y}/${_}`);)_=`${u}-combined-${p}.md`,p++;return _}return t.$$set=u=>{"app"in u&&o(1,i=u.app),"content"in u&&o(2,s=u.content),"processedContent"in u&&o(0,l=u.processedContent),"$$scope"in u&&o(5,n=u.$$scope)},[l,i,s,a,d,n,r]}var it=class extends O{constructor(e){super(),X(this,e,ge,me,q,{app:1,content:2,processedContent:0,processMarkdown:3,getUniqueFileName:4},_e)}get processMarkdown(){return this.$$.ctx[3]}get getUniqueFileName(){return this.$$.ctx[4]}},Ct=it;function ye(t){H(t,"svelte-x3lg6p",".plugin-main.svelte-x3lg6p{width:100%}")}function be(t){let e,o,r=t[4].default,n=B(r,t,t[3],null);return{c(){e=E("div"),n&&n.c(),S(e,"class","plugin-main svelte-x3lg6p")},m(i,s){v(i,e,s),n&&n.m(e,null),o=!0},p(i,[s]){n&&n.p&&(!o||s&8)&&U(n,r,i,i[3],o?z(r,i[3],s,null):G(i[3]),null)},i(i){o||(m(n,i),o=!0)},o(i){F(n,i),o=!1},d(i){i&&$(e),n&&n.d(i)}}}function we(t,e,o){let{$$slots:r={},$$scope:n}=e,{plugin:i}=e,s=null;async function l(){let a=i.app.workspace.getActiveFile();if(!a||!(a instanceof g.TFile)){new g.Notice("Aucun fichier Markdown actif \xE0 combiner.");return}if(a.extension!=="md"){new g.Notice("Le fichier actif n'est pas un fichier Markdown.");return}if(!a.parent){new g.Notice("Le fichier n'a pas de dossier parent.");return}new g.Notice(`Combinaison de ${a.name}...`);let d=await i.app.vault.read(a);s?s.$set({content:d}):s=new Ct({target:document.body,props:{app:i.app,content:d}});let u=await s.processMarkdown(d),y=a.parent.path,p=a.basename,_=await s.getUniqueFileName(p,y),b=`${y}/${_}`;try{await i.app.vault.create(b,u),console.log(`Fichier combin\xE9 cr\xE9\xE9 \xE0 : ${b}`),new g.Notice(`Fichier combin\xE9 cr\xE9\xE9 : ${b}`);let h=i.app.vault.getAbstractFileByPath(b);h instanceof g.TFile&&await i.app.workspace.getLeaf("split","vertical").openFile(h)}catch(h){let T=h instanceof Error?h.message:"Erreur inconnue";console.error(`Erreur lors de la cr\xE9ation du fichier combin\xE9 : ${T}`),new g.Notice(`Erreur lors de la cr\xE9ation du fichier combin\xE9 : ${T}`)}}function f(){i.addCommand({id:"combine-markdown-files",name:"Combiner les fichiers Markdown avec les embeds",callback:l})}function c(){s&&(s.$destroy(),s=null)}return t.$$set=a=>{"plugin"in a&&o(0,i=a.plugin),"$$scope"in a&&o(3,n=a.$$scope)},[i,f,c,n,r]}var st=class extends O{constructor(e){super(),X(this,e,we,be,q,{plugin:0,initialize:1,cleanup:2},ye)}get initialize(){return this.$$.ctx[1]}get cleanup(){return this.$$.ctx[2]}},Mt=st;var K=class extends Ot.Plugin{main=null;async onload(){this.main=new Mt({target:document.body,props:{plugin:this}}),this.main.initialize()}onunload(){this.main&&(this.main.cleanup(),this.main.$destroy(),this.main=null)}};
+"use strict";var P=Object.defineProperty;var ht=Object.getOwnPropertyDescriptor;var mt=Object.getOwnPropertyNames;var xt=Object.prototype.hasOwnProperty;var $t=(t,e)=>{for(var n in e)P(t,n,{get:e[n],enumerable:!0})},gt=(t,e,n,o)=>{if(e&&typeof e=="object"||typeof e=="function")for(let r of mt(e))!xt.call(t,r)&&r!==n&&P(t,r,{get:()=>e[r],enumerable:!(o=ht(e,r))||o.enumerable});return t};var yt=t=>gt(P({},"__esModule",{value:!0}),t);var re={};$t(re,{default:()=>L});module.exports=yt(re);var _t=require("obsidian");function m(){}function bt(t,e){for(let n in e)t[n]=e[n];return t}function R(t){return t()}function M(){return Object.create(null)}function d(t){t.forEach(R)}function A(t){return typeof t=="function"}function V(t,e){return t!=t?e==e:t!==e||t&&typeof t=="object"||typeof t=="function"}function W(t){return Object.keys(t).length===0}function H(t,e,n,o){if(t){let r=J(t,e,n,o);return t[0](r)}}function J(t,e,n,o){return t[1]&&o?bt(n.ctx.slice(),t[1](o(e))):n.ctx}function X(t,e,n,o){if(t[2]&&o){let r=t[2](o(n));if(e.dirty===void 0)return r;if(typeof r=="object"){let i=[],l=Math.max(e.dirty.length,r.length);for(let s=0;s<l;s+=1)i[s]=e.dirty[s]|r[s];return i}return e.dirty|r}return e.dirty}function K(t,e,n,o,r,i){if(r){let l=J(e,n,o,i);t.p(l,r)}}function Y(t){if(t.ctx.length>32){let e=[],n=t.ctx.length/32;for(let o=0;o<n;o++)e[o]=-1;return e}return-1}var I=typeof window<"u"?window:typeof globalThis<"u"?globalThis:global;var D=class t{_listeners="WeakMap"in I?new WeakMap:void 0;_observer=void 0;options;constructor(e){this.options=e}observe(e,n){return this._listeners.set(e,n),this._getObserver().observe(e,this.options),()=>{this._listeners.delete(e),this._observer.unobserve(e)}}_getObserver(){return this._observer??(this._observer=new ResizeObserver(e=>{for(let n of e)t.entries.set(n.target,n),this._listeners.get(n.target)?.(n)}))}};D.entries="WeakMap"in I?new WeakMap:void 0;var Q=!1;function Z(){Q=!0}function tt(){Q=!1}function et(t,e){t.appendChild(e)}function nt(t,e,n){let o=rt(t);if(!o.getElementById(e)){let r=E("style");r.id=e,r.textContent=n,Ft(o,r)}}function rt(t){if(!t)return document;let e=t.getRootNode?t.getRootNode():t.ownerDocument;return e&&e.host?e:t.ownerDocument}function Ft(t,e){return et(t.head||t,e),e.sheet}function k(t,e,n){t.insertBefore(e,n||null)}function x(t){t.parentNode&&t.parentNode.removeChild(t)}function E(t){return document.createElement(t)}function S(t,e,n){n==null?t.removeAttribute(e):t.getAttribute(e)!==n&&t.setAttribute(e,n)}function ot(t){return Array.from(t.childNodes)}function it(t){let e={};return t.childNodes.forEach(n=>{e[n.slot||"default"]=!0}),e}var b;function _(t){b=t}var $=[];var ct=[],v=[],lt=[],Mt=Promise.resolve(),q=!1;function ut(){q||(q=!0,Mt.then(N))}function C(t){v.push(t)}var B=new Set,w=0;function N(){if(w!==0)return;let t=b;do{try{for(;w<$.length;){let e=$[w];w++,_(e),At(e.$$)}}catch(e){throw $.length=0,w=0,e}for(_(null),$.length=0,w=0;ct.length;)ct.pop()();for(let e=0;e<v.length;e+=1){let n=v[e];B.has(n)||(B.add(n),n())}v.length=0}while($.length);for(;lt.length;)lt.pop()();q=!1,B.clear(),_(t)}function At(t){if(t.fragment!==null){t.update(),d(t.before_update);let e=t.dirty;t.dirty=[-1],t.fragment&&t.fragment.p(t.ctx,e),t.after_update.forEach(C)}}function at(t){let e=[],n=[];v.forEach(o=>t.indexOf(o)===-1?e.push(o):n.push(o)),n.forEach(o=>o()),v=e}var T=new Set,Dt;function F(t,e){t&&t.i&&(T.delete(t),t.i(e))}function j(t,e,n,o){if(t&&t.o){if(T.has(t))return;T.add(t),Dt.c.push(()=>{T.delete(t),o&&(n&&t.d(1),o())}),t.o(e)}else o&&o()}var Nt=["allowfullscreen","allowpaymentrequest","async","autofocus","autoplay","checked","controls","default","defer","disabled","formnovalidate","hidden","inert","ismap","loop","multiple","muted","nomodule","novalidate","open","playsinline","readonly","required","reversed","selected"],Tt=new Set([...Nt]);function Pt(t,e,n){let{fragment:o,after_update:r}=t.$$;o&&o.m(e,n),C(()=>{let i=t.$$.on_mount.map(R).filter(A);t.$$.on_destroy?t.$$.on_destroy.push(...i):d(i),t.$$.on_mount=[]}),r.forEach(C)}function Rt(t,e){let n=t.$$;n.fragment!==null&&(at(n.after_update),d(n.on_destroy),n.fragment&&n.fragment.d(e),n.on_destroy=n.fragment=null,n.ctx=[])}function It(t,e){t.$$.dirty[0]===-1&&($.push(t),ut(),t.$$.dirty.fill(0)),t.$$.dirty[e/31|0]|=1<<e%31}function ft(t,e,n,o,r,i,l=null,s=[-1]){let u=b;_(t);let c=t.$$={fragment:null,ctx:[],props:i,update:m,not_equal:r,bound:M(),on_mount:[],on_destroy:[],on_disconnect:[],before_update:[],after_update:[],context:new Map(e.context||(u?u.$$.context:[])),callbacks:M(),dirty:s,skip_bound:!1,root:e.target||u.$$.root};l&&l(c.root);let g=!1;if(c.ctx=n?n(t,e.props||{},(a,y,...h)=>{let p=h.length?h[0]:y;return c.ctx&&r(c.ctx[a],c.ctx[a]=p)&&(!c.skip_bound&&c.bound[a]&&c.bound[a](p),g&&It(t,a)),y}):[],c.update(),g=!0,d(c.before_update),c.fragment=o?o(c.ctx):!1,e.target){if(e.hydrate){Z();let a=ot(e.target);c.fragment&&c.fragment.l(a),a.forEach(x)}else c.fragment&&c.fragment.c();e.intro&&F(t.$$.fragment),Pt(t,e.target,e.anchor),tt(),N()}_(u)}var Bt;typeof HTMLElement=="function"&&(Bt=class extends HTMLElement{$$ctor;$$s;$$c;$$cn=!1;$$d={};$$r=!1;$$p_d={};$$l={};$$l_u=new Map;constructor(t,e,n){super(),this.$$ctor=t,this.$$s=e,n&&this.attachShadow({mode:"open"})}addEventListener(t,e,n){if(this.$$l[t]=this.$$l[t]||[],this.$$l[t].push(e),this.$$c){let o=this.$$c.$on(t,e);this.$$l_u.set(e,o)}super.addEventListener(t,e,n)}removeEventListener(t,e,n){if(super.removeEventListener(t,e,n),this.$$c){let o=this.$$l_u.get(e);o&&(o(),this.$$l_u.delete(e))}if(this.$$l[t]){let o=this.$$l[t].indexOf(e);o>=0&&this.$$l[t].splice(o,1)}}async connectedCallback(){if(this.$$cn=!0,!this.$$c){let t=function(r){return()=>{let i;return{c:function(){i=E("slot"),r!=="default"&&S(i,"name",r)},m:function(u,c){k(u,i,c)},d:function(u){u&&x(i)}}}};if(await Promise.resolve(),!this.$$cn||this.$$c)return;let e={},n=it(this);for(let r of this.$$s)r in n&&(e[r]=[t(r)]);for(let r of this.attributes){let i=this.$$g_p(r.name);i in this.$$d||(this.$$d[i]=z(i,r.value,this.$$p_d,"toProp"))}for(let r in this.$$p_d)!(r in this.$$d)&&this[r]!==void 0&&(this.$$d[r]=this[r],delete this[r]);this.$$c=new this.$$ctor({target:this.shadowRoot||this,props:{...this.$$d,$$slots:e,$$scope:{ctx:[]}}});let o=()=>{this.$$r=!0;for(let r in this.$$p_d)if(this.$$d[r]=this.$$c.$$.ctx[this.$$c.$$.props[r]],this.$$p_d[r].reflect){let i=z(r,this.$$d[r],this.$$p_d,"toAttribute");i==null?this.removeAttribute(this.$$p_d[r].attribute||r):this.setAttribute(this.$$p_d[r].attribute||r,i)}this.$$r=!1};this.$$c.$$.after_update.push(o),o();for(let r in this.$$l)for(let i of this.$$l[r]){let l=this.$$c.$on(r,i);this.$$l_u.set(i,l)}this.$$l={}}}attributeChangedCallback(t,e,n){this.$$r||(t=this.$$g_p(t),this.$$d[t]=z(t,n,this.$$p_d,"toProp"),this.$$c?.$set({[t]:this.$$d[t]}))}disconnectedCallback(){this.$$cn=!1,Promise.resolve().then(()=>{!this.$$cn&&this.$$c&&(this.$$c.$destroy(),this.$$c=void 0)})}$$g_p(t){return Object.keys(this.$$p_d).find(e=>this.$$p_d[e].attribute===t||!this.$$p_d[e].attribute&&e.toLowerCase()===t)||t}});function z(t,e,n,o){let r=n[t]?.type;if(e=r==="Boolean"&&typeof e!="boolean"?e!=null:e,!o||!n[t])return e;if(o==="toAttribute")switch(r){case"Object":case"Array":return e==null?null:JSON.stringify(e);case"Boolean":return e?"":null;case"Number":return e??null;default:return e}else switch(r){case"Object":case"Array":return e&&JSON.parse(e);case"Boolean":return e;case"Number":return e!=null?+e:e;default:return e}}var O=class{$$=void 0;$$set=void 0;$destroy(){Rt(this,1),this.$destroy=m}$on(e,n){if(!A(n))return m;let o=this.$$.callbacks[e]||(this.$$.callbacks[e]=[]);return o.push(n),()=>{let r=o.indexOf(n);r!==-1&&o.splice(r,1)}}$set(e){this.$$set&&!W(e)&&(this.$$.skip_bound=!0,this.$$set(e),this.$$.skip_bound=!1)}};var pt="4";typeof window<"u"&&(window.__svelte||(window.__svelte={v:new Set})).v.add(pt);function Kt(t){nt(t,"svelte-1rhwprr",".combiner-app.svelte-1rhwprr{width:100%}")}function Yt(t){let e,n,o=t[2].default,r=H(o,t,t[1],null);return{c(){e=E("div"),r&&r.c(),S(e,"class","combiner-app svelte-1rhwprr")},m(i,l){k(i,e,l),r&&r.m(e,null),n=!0},p(i,[l]){r&&r.p&&(!n||l&2)&&K(r,o,i,i[1],n?X(o,i[1],l,null):Y(i[1]),null)},i(i){n||(F(r,i),n=!0)},o(i){j(r,i),n=!1},d(i){i&&x(e),r&&r.d(i)}}}async function Qt(t,e){let n=new Set;async function o(r){let i=/!\[\[([^\]#|^]+)(?:(#)([^\]|]+))?(?:(\^)([^\]|]+))?(?:\|([^\]]+))?\]\]/g,l=r,s;for(;(s=i.exec(r))!==null;){let u=s[0],c=s[1].trim(),g=s[2],a=s[3],y=s[4],h=s[5];if(!n.has(c)){n.add(c);try{let p=e.metadataCache.getFirstLinkpathDest(c,"");if(p){let f=await e.vault.read(p);g&&a?f=Zt(f,a):y&&h&&(f=te(f,h));let U=await o(f);l=l.replace(u,U)}else l=l.replace(u,`<!-- Linked file not found: ${c} -->`)}catch{l=l.replace(u,`<!-- Error: Could not process ${c} -->`)}finally{n.delete(c)}}}return l}return await o(t)}function Zt(t,e){let n=t.split(`
+`),o=new RegExp(`^#+\\s*${e}$`),r=!1,i=[],l=0;for(let s of n){let u=s.match(/^(#+)\s*(.*)$/);if(u){let c=u[1].length;o.test(s)?(r=!0,l=c,i.push(s)):r&&c<=l&&(r=!1)}r&&i.push(s)}return i.join(`
+`)}function te(t,e){let n=t.split(`
+`),o=new RegExp(`\\s\\^${e}$`);for(let r=0;r<n.length;r++)if(o.test(n[r]))return n[r].replace(o,"").trim();return""}async function ee(t,e,n){let o=1,r=`${e}-combined.md`;for(;await t.vault.adapter.exists(`${n}/${r}`);)r=`${e}-combined-${o}.md`,o++;return r}function ne(t,e,n){let{$$slots:o={},$$scope:r}=e,{plugin:i}=e;async function l(){let{app:s}=i,u=s.workspace.getActiveFile();if(!u||u.extension!=="md"||!u.parent){s.workspace.trigger("notice","Aucun fichier Markdown valide s\xE9lectionn\xE9.");return}s.workspace.trigger("notice",`Combinaison de ${u.name}...`);let c=await s.vault.read(u),g=await Qt(c,s),a=u.parent.path,y=u.basename,h=await ee(s,y,a),p=`${a}/${h}`;try{await s.vault.create(p,g),s.workspace.trigger("notice",`Fichier combin\xE9 cr\xE9\xE9 : ${p}`);let f=s.vault.getAbstractFileByPath(p);f&&await s.workspace.getLeaf("split","vertical").openFile(f)}catch(f){s.workspace.trigger("notice",`Erreur lors de la cr\xE9ation du fichier combin\xE9 : ${f?.message||f}`)}}return t.$$set=s=>{"plugin"in s&&n(0,i=s.plugin),"$$scope"in s&&n(1,r=s.$$scope)},[i,r,o]}var G=class extends O{constructor(e){super(),ft(this,e,ne,Yt,V,{plugin:0},Kt)}},dt=G;var L=class extends _t.Plugin{appView=null;async onload(){this.appView=new dt({target:document.body,props:{plugin:this}})}onunload(){this.appView?.$destroy(),this.appView=null}};
