--- prompt.md
@@ -1,38 +1,0 @@
-<!-- 
-Repondre en francais 
-affichage  graphique de la structure de des fichers et des dossiers
-
->>> lugin obsidin 
-environnement svelte (version 4et pas plus)
---------------------
-concis , orecis, oratique, intelligent, complet
---------------------
-
-le plugin fonctionne bien, mais j ai besoin que l utilisateur puisse définir lui meme des pattern et couleurs personalisés dans les settigns du plugin (tout doit etre en svelte avannt d etre compilé dans ke main.js)
-1er image  montre un exemple de l ajout par l utilisaateur des patterns,
-2e image, c est une vue actuel des effet du plugin
-
-----------------
- nuulle part pour l utilisateur d entrer couleur et patterns personalisé suplementtaire. ( si je me trompe tu me dis ou trouver
--------------------
- c est paprfait 
- blocs plus compact dans les settings duu plug dans obsidian
- les patterne et couleur de base, actuellemnt caché dans le code doivent etre exposés et modigiable parl l'utilisateur (donc bloc disponible dans les settings avant qu on `+add patern` et relatif au directives dasn fichier venan de sublimetext source d inspiration ) 
- avoir les bloc par defaut 
- voila dans sublimetex ce qu j ai personalisé pour mes colorations syntaxique, peut transformer pour l intégration dans notre plug obsidian?
- la coloration syntaxique 
- ------------------- 
-
- c est compact dans les settings, merci
- voila le settings par defaut, mais je veut plus deataillé . tu dois prendre les contenu regex et couleur pour definir tout comme setting (et l utilisateur n aura plus qu a laisser active ou desactivé certains au besoin) mais toujours avec possiblilité d'ajouter
- je ne sais pas comment sera possible de gerer les imbrications (comme les guillemets entre le parenthesses ou vice-versa etc.)
--->
-
---------------------
-Repondre en francais 
-do:
-- mise a jour en temps reel pendant la saisie dans les settings
-    Il faut actuellement reinitialiser obsidian pour que les changements soient pris en compte.
- 
-
-https://aistudio.google.com/apps/drive/14J_4poXv5YEr8dYz63BNyZnpQpglmyox?showPreview=true
